k256+p256: refactor basepoint tables
Replaces repetitive `cfg` gating by putting them in an
`arithmetic::tables` module, a pattern we can repeat for the other
crates.
