+++ k256/src/arithmetic.rs
@@ -2,14 +2,16 @@
 
 pub(crate) mod affine;
 mod field;
-#[cfg(feature = "hash2curve")]
-mod hash2curve;
 mod mul;
 pub(crate) mod projective;
 pub(crate) mod scalar;
 
 #[cfg(test)]
 mod dev;
+#[cfg(feature = "hash2curve")]
+mod hash2curve;
+#[cfg(feature = "precomputed-tables")]
+mod tables;
 
 pub use field::FieldElement;
