+++ p256/src/arithmetic.rs
@@ -9,33 +9,19 @@ pub(crate) mod scalar;
 
 #[cfg(feature = "hash2curve")]
 mod hash2curve;
+#[cfg(feature = "precomputed-tables")]
+mod tables;
 
 use self::{field::FieldElement, scalar::Scalar};
 use crate::NistP256;
 use elliptic_curve::{CurveArithmetic, PrimeCurveArithmetic};
 use primeorder::{PrimeCurveParams, point_arithmetic};
 
-#[cfg(all(feature = "alloc", feature = "precomputed-tables"))]
-use primeorder::PrimeCurveWithBasepointTableVartime;
-
 /// Elliptic curve point in affine coordinates.
 pub type AffinePoint = primeorder::AffinePoint<NistP256>;
 
 /// Elliptic curve point in projective coordinates.
 pub type ProjectivePoint = primeorder::ProjectivePoint<NistP256>;
-
-/// Window size for the variable-time basepoint table.
-#[cfg(all(feature = "alloc", feature = "precomputed-tables"))]
-const WINDOW_SIZE_VARTIME: usize = 8;
-
-/// Variable-time basepoint table for NIST P-256's generator.
-#[cfg(all(feature = "alloc", feature = "precomputed-tables"))]
-type BasepointTableVartime =
-    elliptic_curve::point::BasepointTableVartime<ProjectivePoint, WINDOW_SIZE_VARTIME>;
-
-/// Lazily computed basepoint table.
-#[cfg(all(feature = "alloc", feature = "precomputed-tables"))]
-static BASEPOINT_TABLE_VARTIME: BasepointTableVartime = BasepointTableVartime::new();
 
 impl CurveArithmetic for NistP256 {
     type AffinePoint = AffinePoint;
@@ -80,11 +66,6 @@ impl PrimeCurveParams for NistP256 {
 
     #[cfg(all(feature = "alloc", feature = "precomputed-tables"))]
     fn mul_by_generator_vartime(k: &Scalar) -> ProjectivePoint {
-        BASEPOINT_TABLE_VARTIME.mul(k)
+        tables::BASEPOINT_TABLE_VARTIME.mul(k)
     }
-}
-
-#[cfg(all(feature = "alloc", feature = "precomputed-tables"))]
-impl PrimeCurveWithBasepointTableVartime<WINDOW_SIZE_VARTIME> for NistP256 {
-    const BASEPOINT_TABLE_VARTIME: &'static BasepointTableVartime = &BASEPOINT_TABLE_VARTIME;
 }
